backend/requirements.txt
python-telegram-bot==20.3
apscheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
base58==2.1.1
anthropic>=0.20.0

pandas>=2.0.0
pandas-ta>=0.3.14
websocket-client==1.6.1
websockets>=12.0
axiom-py>=0.10.0
telethon==1.34.0
solders>=0.21.0
